% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot-index.R
\name{prot.index}
\alias{prot.index}
\title{K-mer indexing function}
\usage{
prot.index(inputFile, w = 0, W = 0, g = 0, G = 0)
}
\arguments{
\item{inputFile}{Path to datafile}
}
\value{

}
\description{
K-mer indexing function
}
\examples{


}

