% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutCorrFilter.R
\name{tutCorrFilter}
\alias{tutCorrFilter}
\title{Correlational filtering}
\usage{
tutCorrFilter(inputFile, c = 0.5, printMat = FALSE)
}
\arguments{
\item{inputFile}{Input k-mer file}
}
\value{

}
\description{
Correlational filtering
}
\examples{


}

