% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutKNN.R
\name{tutKNN}
\alias{tutKNN}
\title{K-Nearest Neighbours}
\usage{
tutKNN(inputFile, knum = 10, t = 33)
}
\arguments{
\item{inputFile}{Input k-mer file}
}
\value{

}
\description{
K-Nearest Neighbours
}
\examples{


}

