% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutKNNcv.R
\name{tutKNNcv}
\alias{tutKNNcv}
\title{Cross validated KNN for analysis
NOT VIEWABLE IN PACKAGE}
\usage{
tutKNNcv(inputFile, folds = 5, paramGrid = expand.grid(k = 2:30),
  repeatn = 3)
}
\value{

}
\description{
Cross validated KNN for analysis
NOT VIEWABLE IN PACKAGE
}
\examples{


}

