% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutNBcv.R
\name{tutNBcv}
\alias{tutNBcv}
\title{Cross validated NB for analysis
NOT VIEWABLE IN PACKAGE}
\usage{
tutNBcv(inputFile, folds = 5)
}
\value{

}
\description{
Cross validated NB for analysis
NOT VIEWABLE IN PACKAGE
}
\examples{


}

