% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutRFE.R
\name{tutRFE}
\alias{tutRFE}
\title{Recursive Feature Elimination}
\usage{
tutRFE(inputFile, featureSize = seq(round(ncol(inputFile) * 0.5),
  round(ncol(inputFile) * 0.25), length.out = 5))
}
\arguments{
\item{inputFile}{Input k-mer file}
}
\value{

}
\description{
Recursive Feature Elimination
}
\examples{


}

