% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutSVM.R
\name{tutSVM}
\alias{tutSVM}
\title{Support Vector Machines}
\usage{
tutSVM(inputFile, kernel = "linear", paramGrid = NULL, gridlen = 20,
  t = 33)
}
\arguments{
\item{inputFile}{Input k-mer file}
}
\value{

}
\description{
Support Vector Machines
}
\examples{


}

