% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAccuracies.R
\name{plotAccuracies}
\alias{plotAccuracies}
\title{Plot accuracies of all methods and their error bars. Used in analysis, not used for tutorial.}
\usage{
plotAccuracies(accuracies, n)
}
\value{

}
\description{
Plot accuracies of all methods and their error bars. Used in analysis, not used for tutorial.
}
\examples{


}

