% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorialAutoAnalysis.R
\name{tutorialAutoAnalysis}
\alias{tutorialAutoAnalysis}
\title{Complete run of all combinations of methods. Used in analysis, not used for tutorial.}
\usage{
tutorialAutoAnalysis(inputFile)
}
\arguments{
\item{inputFile}{Input original file}
}
\value{

}
\description{
Complete run of all combinations of methods. Used in analysis, not used for tutorial.
}
\examples{

}

