% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutRFcv.R
\name{tutRFcv}
\alias{tutRFcv}
\title{Cross validated RF for analysis
NOT VIEWABLE IN PACKAGE}
\usage{
tutRFcv(inputFile, ntree = 100, folds = 5)
}
\value{

}
\description{
Cross validated RF for analysis
NOT VIEWABLE IN PACKAGE
}
\examples{


}

