% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutSVMcv.R
\name{tutSVMcv}
\alias{tutSVMcv}
\title{Cross validated SVM for analysis
NOT VIEWABLE IN PACKAGE}
\usage{
tutSVMcv(inputFile, kernel = "linear", paramGrid = NULL, gridlen = 20,
  folds = 5)
}
\value{

}
\description{
Cross validated SVM for analysis
NOT VIEWABLE IN PACKAGE
}
\examples{


}

